Separated user app from utils to avoid circular imports.
