+++ mmpl_rest/settings.py
@@ -32,7 +32,7 @@ ALLOWED_HOSTS = []
 # Application definition
 
 INSTALLED_APPS = [
-    'utils',
+    'user',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'authtools',
     'djoser',
+    'utils',
     'userprofile'
 ]
 
@@ -128,7 +129,7 @@ STATIC_URL = '/static/'
 # Auth Settings
 
 # Set the custom user model with the email Base User.
-AUTH_USER_MODEL = 'utils.User'
+AUTH_USER_MODEL = 'user.User'
 
 # Rest Framework settings
 REST_FRAMEWORK = {
