+++ utils/admin.py
@@ -1,6 +1,3 @@
-from authtools.admin import NamedUserAdmin
 from django.contrib import admin
 
-from utils.models import User
-
-admin.site.register(User, NamedUserAdmin)
+# Register your models here.
