+++ utils/views.py
@@ -1,7 +1,4 @@
 from django.shortcuts import render
-from django.http import Http404
-
 
 # Create your views here.
-def user_delete_view(request):
-    raise Http404()
+
